generate flutter code examples and add to quickstart docs
