+++ examples/flutter/test/quick_start_test.dart
@@ -4,18 +4,6 @@ import 'package:provider_shopper/models/car.dart';
 import 'package:realm/realm.dart';
 
 void main() {
-  group('Open Realm', () {
-    test('open a Realm', () {
-      // :snippet-start: open-realm
-      var config = Configuration();
-      config.schema.add(Car);
-      var realm = Realm(config);
-      // :snippet-end:
-      expect(realm.isClosed, false);
-      realm.close();
-    });
-  });
-
   group('CRUD Operations', () {
     var config = Configuration();
     config.schema.add(Car);
