+++ source/sdk/flutter/quick-start.txt
@@ -75,12 +75,8 @@ would like to open, including schema.
 Pass your configuration to the `Realm constructor <https://pub.dev/documentation/realm/latest/realm/Realm-class.html>`__
 to generate an instance of that {+realm+}:
 
-.. code-block:: dart 
-
-   var config = Configuration();
-   config.schema.add(Car);
-
-   var realm = Realm(config);
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.open-realm.dart
+   :language: dart
 
 You can now use that {+realm+} instance to work with objects in the database.
 
@@ -96,11 +92,8 @@ Create Objects
 To create a new ``Car``, instantiate an instance of the
 ``Car`` class and add it to the {+realm+} in a write transaction block:
 
-.. code-block:: dart
-   
-   realm.write(() {
-     var car = realm.create(Car()..make = "Tesla"..model = "Model Y"..kilometers = 42);
-   });
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.create-realm-object.dart
+   :language: dart
 
 Update Objects
 ~~~~~~~~~~~~~~
@@ -109,9 +102,8 @@ To modify a car, update its properties in a write transaction block:
 
 .. TODO(DOCSP-20150): couldn't find anything in the current docs
 
-.. code-block:: dart
-
-   // find out how to update
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.update-realm-object.dart
+   :language: dart
 
 Query for Objects
 ~~~~~~~~~~~~~~~~~
@@ -120,18 +112,14 @@ Retrieve a collection of all objects of a data model in the {+realm+} with the
 `Realm.objects<RealmObject>() <https://pub.dev/documentation/realm/latest/realm/Realm/objects.html>`__
 method:
 
-.. code-block:: dart
-   
-   var cars = realm.objects<Car>();
-   Car myCar = objects[0];
-   print("My car is ${myCar.make} ${myCar.model}");
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.query-all-realm-objects.dart
+   :language: dart
 
 You can also filter a collection to retrieve a more specific collection
 of objects:
 
-.. code-block:: dart
-
-   cars = realm.objects<Car>().where("make == 'Tesla'");
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.query-realm-objects-with-filter.dart
+   :language: dart
 
 .. TODO(DOCSP-20150): add something about using sort  
 
@@ -148,20 +136,14 @@ Delete Objects
 Delete a car by calling the `Realm.delete(RealmObject) <https://pub.dev/documentation/realm/latest/realm/Realm/delete.html>`__
 method in a write transaction block:
 
-.. code-block:: dart
-
-   realm.write(() {
-     realm.delete(myCar);
-   });
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.delete-one-realm-object.dart
+   :language: dart
 
 Delete multiple cars with the `Realm.deleteMany(List<RealmObject>) 
 <https://pub.dev/documentation/realm/latest/realm/Realm/deleteMany.html>`__ method in a write transaction block. 
 
-.. code-block:: dart
-
-   realm.write(() {
-     realm.deleteMany(cars);
-   });
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.delete-many-realm-objects.dart
+   :language: dart
 
 Listen for Changes to Objects
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -173,21 +155,18 @@ Listen and respond to changes to an entire {+realm+} database, a collection or a
 
 To listen to an entire {+realm+}, use the `Realm.addListener(Event, ListenerCallback) <https://pub.dev/documentation/realm/latest/realm/Realm/addListener.html>`__ method. 
 
-.. code-block:: dart
-
-   // TODO(DOCSP-20150): add code example
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.listen-to-entire-realm.dart
+   :language: dart
 
 To listen to an a collection of {+realm+} objects, use the `Results<RealmObject>.addListener(ResultsListenerCallback) <https://pub.dev/documentation/realm/latest/realm/Results/addListener.html>`__ method. 
 
-.. code-block:: dart
-
-   // TODO(DOCSP-20150): add code example
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.listen-to-collection-realm-objects.dart
+   :language: dart
 
 To listen to an a single {+realm+} object, use the `RealmObject.addListener(ResultsListenerCallback) <https://pub.dev/documentation/realm/latest/realm/RealmObject/addListener.html>`__ method. 
 
-.. code-block:: dart
-
-   // TODO(DOCSP-20150): add code example
+.. literalinclude:: /examples/generated/flutter/open_realm_test.codeblock.listen-to-single-realm-object.dart
+   :language: dart
 
 Complete Example
 ----------------
